+++ be/airbnb/src/main/java/com/team19/airbnb/repository/UserRepository.java
@@ -1,0 +1,4 @@
+package com.team19.airbnb.repository;
+
+public interface UserRepository {
+}
