+++ be/airbnb/src/main/java/com/team19/airbnb/repository/WishlistRepository.java
@@ -1,0 +1,4 @@
+package com.team19.airbnb.repository;
+
+public interface WishlistRepository {
+}
